+++ rai-sdk/src/main/java/com/relationalai/TransactionAsyncCancelResponse.java
@@ -1,0 +1,8 @@
+package com.relationalai;
+
+import com.jsoniter.annotation.JsonProperty;
+
+public class TransactionAsyncCancelResponse extends Entity {
+    @JsonProperty(value = "message")
+    public String message;
+}